- fix travis - split validator from index file - rename main to index - use same json-schema validator as with schulcloud-content
